Fix the "Ghost minecart" issue.
This reverts commit 73647864179cbab9f340044e9d33add445fbe774 - "Ignore entity movement packets if distance == 0"

This a commit introduced an intermittent bug which has caused minecart ghosting - they are invisible until you interact with them physically. This is described in leaky bugs:
http://leaky.bukkit.org/issues/547
http://leaky.bukkit.org/issues/563

And caused by the fact that normally when loading a chunk, entities fall with gravity before the block data has been loaded, and are sent a zero-distance packet to reset their positions. See Issue #190:
https://github.com/Bukkit/CraftBukkit/issues/#issue/190