Allow inventory creation by InventoryType and title. Fixes BUKKIT-4045
Up until now it has not been possible to create a new Inventory using
a custom title and permit any InventoryType available.

The commit changes that by adding a method to optionally supply the title
for the given inventory type and holder, creating the functionality to
display any supported inventory type with a 32 character length String.

If the inventory title supplied is larger than 32 characters then an
IllegalArgumentException is thrown stating so.